requirements.txt
openai-whisper
llama-cpp-python
python-dotenv
google-generativeai
yt-dlp
youtube-transcript-api
PyMuPDF
whispercpp
